Make details of Armor parsing available to consumers
This adds an Armor struct that represents an ASCII-armoring of a packet
stream. It includes the armor type, as well as any headers parsed out of
the armored packet.

This also cleans up the public interface a little bit, importing any
interface structs into the top level namespace - although not
necessarily the details of those structs. This means, for example, that

    pgp::PublicKey

is available, but if you want to match on the variants of that, you'll
need to import them from pgp::publickey first.