Make PacketIterator the public interface
the read_stream function didn't do much, so I'm ripping it out.

In addition, deciding whether the results of a PacketIterator should be
collected into a Vec<Result> or a Result<Vec> is probably
application-specific. By making the iterator the standard interface,
we no longer imply that either of those is "correct"